3.2.1.0
Attempted to fix interaction between Execute and Ricochet, couldn't come up with good tests for that. Not merging with master branch until I'm sure that that's how it works. (Currently if execute is triggered, it overwrites the damage, even if that damage was more than enough to kill a unit, according to replays)